conf.py: Support some arguments
Callers can now run this little programlet (is that a word?) like so:

  conf.py objects.cfg:hostgroup1,hostgroup2

in order to get a list of objects somehow associated with hostgroup1
and hostgroup2 written to the file objects.cfg.

It appears to fsck some things up right now though (host exclusion in
services, for example, such that *,!somehost doesn't exclude somehost
from the services output).

This is one key part to fixing the unsubmitted issue of configuration
sync between a master and a poller.

Signed-off-by: Andreas Ericsson <ae@op5.se>