Add the merlin-apps package and revamp the specfile majorly
Since we introduce the merlin-apps package (replacing the not
publicly released monitor-distributed package), we take the
chance to do a major overhaul of the specfile while we're at it.

The merlin-apps package contains a bunch of free-standing apps
and the general-purpose wrapper 'mon' which takes care of
running them for people.

I'd be surprised and delighted if this package builds now.

Signed-off-by: Andreas Ericsson <ae@op5.se>