mon: Run python scripts without invoking the shell
Since we can invoke python scripts as modules rather than as
standalone scripts and thereby gain not only performance benefits
but also a lot of saved coding effort, we should be able to do that
using the mon wrapper.

This patch adds just such functionality. If a python script resides
in the libexec directory, it's scanned for cmd_* functions and those
functions are then added as commands to the category the script is
supposed to handle. If no such functions exist, or if the module
has a global variable named "pure_script" which is set to True (or
any non-false value), we still treat it as a standalone script.

This means standalone scripts in the libexec directory must use the
if __name__ == '__main__': trick if they're freestanding, but that
shouldn't be much of a problem.

Signed-off-by: Andreas Ericsson <ae@op5.se>