Introduce C-oconf
This provides a small helper application that runs the same routines
as merlin does internally. It can print the sha1 checksum of the current
configuration and print the time it was last changed. The last-changed
thing could easily have been hacked up in python, but this way meant
basically that we only had to add a few more lines of code to reuse the
C routines to do the same job, with less code to maintain.

Signed-off-by: Andreas Ericsson <ae@op5.se>