daemon: Make path to merlin config file-scoped
And also rename it to merlin_conf to avoid clashes with the module's
"config_file" variable exported from Nagios.

This is handy for when we need to split the configuration, since the
mon helper doesn't necessarily know where to find the merlin config,
and it needs to know about it in order to be able to parse it to get
to know where to send the configuration once it's split.

Signed-off-by: Andreas Ericsson <ae@op5.se>