daemon: Change how we handle helper applications
This patch moves a bit of code from the import_objects_and_status()
functions to run_import_program() (now renamed to run_program()),
making the latter handle fork()'ing and things like that. This is in
order to make it a bit more generic and prepare it for handling the
configuration synchronization scripts that we'll want to run when we
detect a configuration change.

Since this change also opens up for more children than one running at
the same time, we modify the child reaping code and its caller
accordingly.

Signed-off-by: Andreas Ericsson <ae@op5.se>