Introduce oconf.push.sh
This is a small, small script that rests on the shoulders of giants.
We get away with some really sparse code since we know that it
shouldn't get run unless the user explicitly asks us to, or unless
the merlin daemon has noticed that we have a newer configuration
than at least one of our peers (or possibly pollers).

This is the last piece of the puzzle to fix issue #3420.

Signed-off-by: Andreas Ericsson <ae@op5.se>