daemon: Don't decrement num_children unless we reaped one
The check for pid == 0 was missing in the code, causing us to
prematurely decrement the num_children counter and thus never
reap the children we had running.

This caused defunct processes and, more importantly, meant we
never marked the import program as having finished and thus
never spawned a new one on configuration changes.

This patch fixes that, making configuration imports work properly
once again.

Signed-off-by: Andreas Ericsson <ae@op5.se>