oconf push app: Make sure only one instance at the time is running
The lockfile gracetime is set to 300 seconds, so if the script
crashes (how the hell...) and doesn't remove its lockfile we
should be able to push configuration shortly afterwards.

The --force switch can be used to forcibly remove the lockfile and
cause a push to happen no matter how many other instances are
running concurrently.

Signed-off-by: Andreas Ericsson <ae@op5.se>