Make merlin (optionally) store performance data
There are graphing tools available that uses performance data
stashed in a relational database. This patch makes Merlin a
usable backend system for such graphing solutions. It's not
enabled by default since we won't be using it on op5 systems,
but at least it's easily enabled for those who want such a
feature now.

Considering how easy it was to hack up, there's no real reason
why we shouldn't support this out of the box.

Signed-off-by: Andreas Ericsson <ae@op5.se>