oconf helper app: Use one variable for nagios.cfg path
We used to have two; ncfg_path and nagios_cfg. I'm not sure which
one was used most, so I just left the one I found first and removed
the other one.

While we're at it, we remove the 'global' declarations of nagios_cfg
in functions that only read from it. It's only needed on assignment
(yes, very odd indeed).

Signed-off-by: Andreas Ericsson <ae@op5.se>