modules/compound_config: Save path and start and end line nr
It's useful for error-reporting, and also in case we want to remove
nodes from the merlin.conf file without touching the comments. For
that purpose, we can use a simple

  sed -i <start>,<end>d path

to remove a node, and once we can remove and add nodes, we can easily
rename and modify too (well, delete and recreate really, but since
there's so very little data it amounts to the same thing, more or less).

Signed-off-by: Andreas Ericsson <ae@op5.se>