mon node: Add the 'status' command to help debug merlin installations
Currently it ignores all arguments and prints detailed status output
for all nodes that were configured when Merlin was started. Problematic
figures or values, such as high latency, node inactiveness or failure
to send frequent enough program_status events are reported in red
color on the terminal. It does not affect the exit-status of Merlin
itself though.


Later, the helper can be plugin-ified, but for now this works nicely
enough.

Support for mysql and postgresql is already available, assuming the
proper python bindings are installed. On CentOS, the packages are
called MySQLdb-python and postgresql-python, respectively.

Signed-off-by: Andreas Ericsson <ae@op5.se>