oconf push: Fix hostgroup parsing for pollers
If a poller has no hostgroups configured, we need to bail out
early. Merlin wouldn't accept such a configuration anyways, so
it can't be the active configuration.

We also modify the call to 'mon node show' slightly to not use
eval, as the node helper doesn't add shell-compatible syntax
and there's no sane way we can do it on our own that's cheaper
than just using sed to get the right value with a simple
assignment anyway.

While we're at it, we get rid of the ADDRESS variable that we
didn't use anyway and quote each argument to mon oconf split
so that hostgroup lines with spaces will work properly.

Signed-off-by: Andreas Ericsson <ae@op5.se>