mon helper: Let command-module's init functions manipulate arguments
Since we changed it so that only the init function of the command
module we'll be running a command from later is run, we can easily
let the init function modify the command arguments.

This is nifty since it lets a single function parse all those
variables available to (very nearly) all the commands in the
category, withouth each command having to care about ignoring them
when they parse their arguments.

Signed-off-by: Andreas Ericsson <ae@op5.se>