Break out the merlin.conf parser to a separate helper module
The command modules are starting to interact quite a bit, so
it's nifty to let them share the common code rather than
duplicating it or resorting to parsing the output of other
commands.

The node command module uses it from the start and provides
a nifty way of testing the new module to make sure it works
as the old code used to.

Signed-off-by: Andreas Ericsson <ae@op5.se>