mon: Let command modules modify the arguments for its commands
Command modules want to parse certain common arguments used for
its commands and then have the commands themselves not bother
about them. This patch makes that possible by checking the
return type from the modules init function and replacing the
args passed to the command itself with the returned value if
that value is a list.

This seems to work well in both theory and practice.

Signed-off-by: Andreas Ericsson <ae@op5.se>