misc.c: Handle cfg_dir options properly
Previously we used to silently skip a lot of files if the first
object config file specifying entry in nagios.cfg was a cfg_dir
statement. This patch corrects that by altering the direction
the linked list grows. New entries are added to the "head" of
it and has its 'next' pointing to the previously added entry in
the list. This saves us a lot of headache, since the list is
always NULL when we start and the first entry thus gets the
NULL sentinel for free from the start.

Signed-off-by: Andreas Ericsson <ae@op5.se>