module: Prevent being loaded into non-matching object structure layout
Nagios provides the nifty variable __nagios_object_structure_version,
which is a versioning number for the object structures (well, doh!)
set at compile time.

The headers we ship for Nagios have a defined version which decides
what offsets to use when coding and decoding packets for node-to-node
transport. If those two versions differ, coding and decoding will
almost certainly not work properly. In such cases, the only sane
way left for us to take is to abort early and give the user a
descriptive message stating why the module fails to load and how to
fix it so that it does.

This will hopefully shed some light of the problems experienced by
some users which experience coding/decoding issues without obvious
reasons.

Signed-off-by: Andreas Ericsson <ae@op5.se>