Mark non-new info with a special return code
Since we for the most cases don't really need to do anything when
the info we get back on a CTRL_ACTIVE packet isn't new, we might
as well mark such occasions with a special return code. Since all
the negative ones are taken, we stick with '1' for this special
case. The special cases are really starting to pile up...

Signed-off-by: Andreas Ericsson <ae@op5.se>