daemon: Stop falling back to using the random hash_delta for csync
The hash delta could be just about anything. Since the config sync
can wipe usercreated configurations when used, we avoid any action
at all when we can't determine which direction is the right one.
Instead we escalate the log-message to a warning, which means it
should get logged for pretty much everyone.

Signed-off-by: Andreas Ericsson <ae@op5.se>