daemon: Reduce waiting time when reading input from the network
When Merlin is connected to more than one node and all nodes are
quite busy, we could otherwise spend a lot of time in poll() to
see if there's input available, which means we could ignore other
nodes which have input waiting while new input arrives on the node
we just read from.

This can still happen, but with this patch, the chance of that
happening is reduced quite a lot.

Signed-off-by: Andreas Ericsson <ae@op5.se>