Fix some cross-wordsize compilation printf() format warnings
We remove many uint64_t references, which seem to lack a sensible
way to be represented on architectures with different wordsizes.

Since we also prefer not to use the (non-portable) %zu format char,
which is mainly used for sizeof() printf()'s, we turn sizeof() into
a macro and cast it to an unsigned int, which is portably printable
as %u.

Signed-off-by: Andreas Ericsson <ae@op5.se>