module: Let Nagios re-schedule the checks we block
When we override a service-check with the NEBERROR_CALLBACKOVERRIDE
return code, Nagios doesn't reschedule the check and so it will
never run again in case we happen to switch peer_id's or if a
poller suddenly goes offline.

Returning NEBERROR_CALLBACKCANCEL causes Nagios to reschedule the
service check so we can block it again the next time it's called.

From a performance-standpoint, it would probably be better to
use NEBERROR_CALLBACKOVERRIDE and then manually reschedule the
service-checks when we discover that a peer or poller changes
state, but we'll try this (much simpler) version first and see how
it pans out.

Signed-off-by: Andreas Ericsson <ae@op5.se>