module: Avoid infinite looping when reaping ipc events
The io cache thing seems to be a fairly huge success, but when we
hit the bufsize mark exactly and ioc->offset and ioc->buflen was
the same, we would enter an infinite loop, eating nearly 100% cpu
and basically just hogging the system entirely.

This patch corrects that particular problem. Stress-testing is so
nice.

Signed-off-by: Andreas Ericsson <ae@op5.se>