Add node_send() to send(2) data to nodes
Just as we have node_recv() to unify error handling when reading from
nodes, we introduce node_send() to not have to duplicate error handling
in the normal path (where events get sent immediately) and in the
binlog sending path, where events may or may not get sent properly.

A bug in the binlog sending code caused the node to not be marked as
disconnected even though we received EPIPE when trying to write to it.
This didn't really cause any grievous harm, but it did mean that we
read and unread the binlog quite a lot when it wasn't really necessary,
and that we spammed the logs quite a lot (one failed send() for each
binlog entry).

Signed-off-by: Andreas Ericsson <ae@op5.se>