module: Reap continuously while we reap a full buffer
Although we do put a max cap on the reaping at 10 loops and
issue a warning when we've looped 3 times, we must make sure
to keep the buffer as empty as we possibly can in order to
avoid congestion in the network. People with HUGE networks
simply shouldn't use forwarding-nodes or top-level masters
to do any real work.

Signed-off-by: Andreas Ericsson <ae@op5.se>