Remove some dead code
The code we're now removing hasn't been used since we moved to the
reading events in bulk, so we might as well delete it. ipc_read_event()
was the last caller of node_read_event(), which in turn was the
only remaining caller of io_recv_all(). Since nothing any longer
calls ipc_read_event(), we can get rid of them all.

Signed-off-by: Andreas Ericsson <ae@op5.se>