module: Use a separate thread to reap events
Though I'm loathe to admit it, using a second thread only to reap
removes the only real bottleneck currently in the system. Before
this patch, we struggled mightily to handle 150k checks / 5 mins.
With this patch, we're handling 674760 checks / 5 mins without
even breaking a sweat.

Fortunately, every API in merlin is already threadsafe, so the
effort to add this patch is minimal, to say the least.

Signed-off-by: Andreas Ericsson <ae@op5.se>