module: Avoid sending superfluous host and service status data events
It's necessary to send these events when they are triggered by a
command causing them to enter or exit downtime, or when they're
being rescheduled, or... Well, actually it's necessary to send
them in a lot of circumstances, but the main source of these
events is when checks are being run in the normal fashion.

It *seems* as though the event chaing goes like this:
* Nagios decides to run a check, so it creates an (A)SYNC_PRECHECK
  event.
* It then immediately procedes to set the check as "is_executing",
  which triggers a status data event.
* After this comes a random period of time while the check is
  running. During this time, other events happen.
* When the check is done, a check result data event is sent to
  the module, containing the next_check time and all that jazzy
  stuff.
* When the check result eventhandler returns, a status data
  event is triggered, containing the exact same data (except
  it's not exactly the same, since it's now a different type
  of event).

If we instead block the check, we get one event immediately
after having done so.

The quick way out of sending redundant events is ofcourse to
cache the object we're about to check and which has just
been checked and see if the next status data event matches
the object we just received a note about.

We use a silly short horizon of 1 second to make sure we never
drop status data events that should be sent.

The long way out, which is tested in case the short way fails,
is to check if a poller or peer is responsible for this check
and, if so, ignore sending the event even if it's not directly
related to a host or service check.

This should ensure that host and service status data events
only ever come from a node that isn't checking it at the same
time, and will save a considerable amount of stress from the
database.

Signed-off-by: Andreas Ericsson <ae@op5.se>