module: Update host and service objects atomically
Using the old code, we would sometimes assign NULL to the various
check output variables of host and service objects, which would
remain in place for a short while while the objects where updated.

With this patch, we try for atomicity in that there should always
be a value set in the various check output pointers that isn't
NULL.

Hopefully this will solve the crashes we've seen when a receiving
module takes care of performance-data from a remote node.

Signed-off-by: Andreas Ericsson <ae@op5.se>