module: Try hard to get proper values for send_paths()
It seems as if Nagios strips the macros we need of their values
when it's given the command to reload its configuration through
the command pipe. To counter that, we first attempt to read the
macro from the config file and, if that fails, we get it from
when Nagios first started, and also copy it back to Nagios when
that happens.

Signed-off-by: Andreas Ericsson <ae@op5.se>