module: free() some readily available memory when deiniting
In case we're doing a soft reload we would otherwise leak that
memory. There are still some minor leaks in the module upon
such events, so the recommended way of restarting Nagios is
to use a hard restart (ie, stop, then start anew) which means
the kernel takes care of releasing all the memory for us.

Signed-off-by: Andreas Ericsson <ae@op5.se>