module: sleep() in short intervals when stalling
While it's good to avoid pegging the CPU, it's not quite so good to
always sleep() until the stall-timer expires. This patch changes it
so we usleep(500) instead, which is plenty of time for the cpu to
relax but so short that the time between the CTRL_RESUME packet
arriving and the stall lock expiring looks and feels instantaneous.

Signed-off-by: Andreas Ericsson <ae@op5.se>