daemon: Attempt to reinit ipc if module goes AWOL for a while
Right now we have no safety catches in place in case anyone
deletes the socket in the filesystem and the module disconnects.
This patch adds one by the simple expedient of attempting to
re-initialize the socket if the module isn't connected and we
haven't tried reinitializing in a while (5 seconds, hardcoded).

Signed-off-by: Andreas Ericsson <ae@op5.se>