sql: Plug a memory leak
When passed an illegal variable, we used to leak the copied value
in the error path. This patch plugs it.

free() is supposed to gracefully accept NULL, but on some systems
that's not the case, so we play it safe and check that value_cpy
is set before free()'ing it.

Reported-by: Stephan Beal <sgbeal@googlemail.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>