Compile without errors on *BSD systems
BSD doesn't have __WORDSIZE, and __BYTE_ORDER is called _BYTE_ORDER
in BSD land. This patch adds support for __WORDSIZE on all machines
where Merlin is likely to be ever considered to be compiled, and if
neither __BYTE_ORDER nor _BYTE_ORDER are defined, we will bomb out
during compilation with a sane-ish error message.

Signed-off-by: Ken Menzel <Ken.Menzel@fisglobal.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>