Disable crontab entries by default
If we start running imports in the middle of the night while
Merlin is still running, we'll end up blocking database access
for the daemon, which is obviously a Bad Thing(tm).

This patch provides instructions on how to enable the cronjobs
in a safe manner for those who want to for whatever reason.

While we're at it, we change logdata import to be incremental,
which should save about a billion cycles on large installations.
Since logfiles are parsed before we get the latest timestamp
from the database, we should basically always end up with a
no-op action in case someone enables this by mistake.

Signed-off-by: Andreas Ericsson <ae@op5.se>