oconf nodesplit: Print friendlier message when no pollers configured
Previously we printed "No pollers configured, so nothing to do"
but the user wasn't aware that this only pertained to splitting
the configuration.

This patch changes the wording so it becomes
"No pollers configured. No need to split config" instead, which
makes it look nicer when called from 'mon oconf push' in case
one runs it manually.

Suggested-by: Ola Sandström <ola.sandstrom@WSPGroup.se>
Signed-off-by: Andreas Ericsson <ae@op5.se>