oconf: Set mtime of created files to same as its source
When a master with peered pollers generates config for those pollers,
it usually takes a few seconds to do so, causing the timestamp on the
pollers' generated config files to be off by a few seconds from that
of the masters. This can lead to ping-pong running of the push command,
since each poller will want to push its configuration to the master
server. Making the timestamps equal for all config files alleviates
this issue in the cheapest possible way.

Signed-off-by: Andreas Ericsson <ae@op5.se>