oconf: Use exact timestamps when determining if config is cached
Since we're setting the mtime of newly created files to the exact
same as that of our current configuration, we should make sure
that the configuration timestamp matches exactly before generating
files.

If we don't do this, we won't re-create configuration files that are
spawned from a restored backup, for example.

Signed-off-by: Andreas Ericsson <ae@op5.se>