db schema: Create indices for the host/service -> group junctions
This decreases query times for host- and service group pages for large
installations. For small installations, or with few such groups, it
weirdly enough causes a small increase in query time. Since small
installations generally return their query results in less wall-clock
time than is humanly noticeable anyway, we don't really care about
those for this particular case.

The current performance cutoff is not known, so there's no telling how
large the installation needs to be for this to be significant. On a
config with 300 hostgroups, 10000 hosts and 92000 services, it provides
nearly 50% reduction in query times for some long-running queries
though.