sql: Don't rerun queries on missing tables
There's a whole slew of occasions where we shouldn't even attempt
to reconnect and rerun the query. This just happens to be one of
them that I've run into. Since it's currently the only way to
disable logging report_data, we support it properly without adding
massive amounts of data to the logs, or leaking memory (as libdbi
does when a new connection is established).

Signed-off-by: Andreas Ericsson <ae@op5.se>