merlin_conf: Check for task assignment conflicts when parsing nodes
If a poller node is assigned the hostgroups A, B and C, and another
poller is assigned the hostgroups B, C and D, that will now result
in both the nodes being marked with conflicting assignments.

It's up to potential callers to display such conflicts, but at least
the info is there now.

Note that this patch does nothing to handle the case when a host (X) is
a member of hostgroups A and B while those hostgroups are assigned to
different pollers. In order to determine that, we need to access the
data in a populated database (or parse the config, but that's a real
hassle, basically).

While we're at it, we also count how many peers each poller has so we
deterministically can decide exactly how many hosts and services each
poller should be checking.

Signed-off-by: Andreas Ericsson <ae@op5.se>