merlin_conf: Add in_config and in_db variables to the merlin_node class
These are handy when displaying status, in case we attempt to display
the node status while the on-disk config differs from the running one,
which would definitely cause confusion. Nodes that we read in from the
config files are marked as being in_config. We never touch in_db,
except in the constructor where we set it to False.

Signed-off-by: Andreas Ericsson <ae@op5.se>