module: pthread_cancel() before pthread_join()
pthread_join() waits while the other thread finishes, so in most
normal cases we would get a segfault if we try to join the thread
that isn't running.

On fedora 12, this seems to be patched in glibc. On CentOS 5.5 (and
thus RHEL5), it seems this is not the case, and a segfault occurs
when attempting to cancel a thread that isn't running.

Signed-off-by: Andreas Ericsson <ae@op5.se>