module: Optionally skip pthread_cancel()'ing the reaper thread
Given the relatively large amount of crashes we've seen from this
call and the fact that the reaper kills itself when cancel_reaping
is set to 1 (which is done in nebmodule_deinit()), this gives us
the option of using both belts and suspenders (default), or just
ignore whichever of the two pthread_cancel() happens to represent.

The module section of the config now groks "cancel_threads = <bool>"
to control this behaviour.

This provides a backway out of issue #3883.

Signed-off-by: Andreas Ericsson <ae@op5.se>