module: Send NULL as second argument to pthread_join()
We don't really care about the return code of the thread anyway,
so we might as well save the thread library some effort. Given
the backtrace of recent coredumps it also seems that pthread_join()
with an uninitialized variable causes issues with some pthread
implementations, even though the glibc source, and especially its
tests, seems to indicate this shouldn't be the case.

Signed-off-by: Andreas Ericsson <ae@op5.se>