daemon: Don't spam the logs with connection attempts
When a collaborating node goes to lunch we used to spam the logs
for every connection attempt we made, which can quite easily be
several times per second.

This patch adds the same logic (and gracetime) as the ipc connection
thing has had for a long time, blocking the logging of the attempt
but not the attempt itself if less than 30 seconds has passed since
we last tried to connect to the remote node.

We also update the ipc stuff to use the per-node variable instead
of a file-scoped one.

Signed-off-by: Andreas Ericsson <ae@op5.se>