example.conf: Demonstrate the cancel_threads variable
We have it, so we might as well showcase it. It's quite likely to go
away once the pthread_{join,cancel}() crap has been fully sorted out
so that it works well on all systems, but for now this is a nifty way
to test multiple patches with a single config switch.

Signed-off-by: Andreas Ericsson <ae@op5.se>