showlog: Support listing files
The import program already has this feature, but it lacks the --first
and --last arguments which causes the program to automatically find
the files within the specified timelimit. Since it's quite easy to
add this support and we'll need it when we implement the log-wiping
helper, we add this simple piece of code.

Signed-off-by: Andreas Ericsson <ae@op5.se>