showlog: Hide uninteresting eventhandler events
"Uninteresting" is used broadly here. It can also refer to items
the user isn't allowed to see. This enables the same type of
filtering that we use for downtime and flapping events for hosts
and services. Since those two event-types seem to work ok, it's
safe to assume that this will too.

This fixes issue #3885.

Signed-off-by: Andreas Ericsson <ae@op5.se>