mtest: Let module code handle connecting and stalling
Function tests are supposed to use a holistic approach, so we should
try to run as high-level stuff as possible and make sure the designated
entrypoints arrive at the proper exitpoints instead of hopping in
somewhere in the middle and cheat our way out of half the course.

Incidentally, this also makes the mtest program continue when it
receives the CTRL_RESUME control packet instead of, as before, stay
dormant for the whole 20 seconds, making tests a bit more pleasant to
run.

This also means we're now testing the thread cancellation and joining,
which we didn't do before. Given the recent bugs in that particular
area, that's most certainly a Good Thing(tm).

Signed-off-by: Andreas Ericsson <ae@op5.se>