mtest: More tests
Specifically, we now also test the dupe-remover, perfdata handling
(if configured), and make sure report_data tests work fine even if
we run the test-case multiple times against the same daemon. This
also makes sure we test the duplicate-state ignoring for report
data in the daemon's database updater.

In addition with the previous changes to the tests, which added
checks for the thread creation and reaping stability, we can now
be fairly safe that this version works as it should for all more
or less normal cases. A cut-n-paste error while writing the new
tests also managed to check that the daemon is fairly retard-
resistant in what kind of data it swallows.

Signed-off-by: Andreas Ericsson <ae@op5.se>