import: Make failures in parsing downtime commands non-fatal
We use downtime commands in a rough and ready hack and slash kind of
way, allowing it to give us hints about what downtime, if any, is
scheduled, and then using that information to make an educated guess
at the real downtime start, stop and duration times. We can make do
without it though, which is proven by this patch.

Signed-off-by: Andreas Ericsson <ae@op5.se>