Fix thinko in event reception code
Previously we used to use the read-offset as the marker for how much
data we were to read, but later in the code we'd proceed and use the
write-offset as the write destination (as is proper). This caused a
certain amount of digestion in the reception IO cache, which was only
frequently visible with fairly large packets but could happen for smaller
packets as well, provided enough of them pile up without a valid header
at the start (which we couldn't, and still can't, detect).

Eventually this lead to both daemon and module crashing out at the same
time. The module because it tries to access the data sent by the daemon,
and the daemon because it tries to access the data it has already sent
to the module when it inserts it into the database.

Tricky one to chase down, so we'd better safeguard against it in the
future. More patches to follow.

Signed-off-by: Andreas Ericsson <ae@op5.se>