test_utils: Make TEST_PASS a positive integer
Since the testing functions are called 'ok_<vartype>', it looks
kinda funny when one wants to abort based on test-results and get
to write

  if (ok_int(1, 2, "random crap-test"))
    exit(EXIT_FAILURE);

since it sort of looks as if we're exiting if the test succeeds.
This swaps the meanings around so we get to properly negate the
situation above, thus getting a much more intuitively readable
code.

Signed-off-by: Andreas Ericsson <ae@op5.se>