Add packet signature to the protocol header
This gives receiving nodes a fighting chance to make sure the packet
received is what the sending end actually sent. A packet that doesn't
have the merlin event signature at the start will get discarded and
the offending sender node disconnected.

This is most likely how the protocol should have been implemented
from the start.

This would require a protocol version bump, were it not for the fact
that the protocol marker of too old merlins now have the wrong offset
for to even see it. The signature will make sure older merlins barf
properly on packets it doesn't understand though, so the protocol
version gets reset to 0 even before it gets bumped.

It does require a distinct version change though, so merlin 1.0.0,
here we go.

Signed-off-by: Andreas Ericsson <ae@op5.se>