mtest: Test protocol stability using packets with malformed length
The receiving end should disconnect us quite a lot, since every other
packet we send with a bad offset address will not have the proper event
signature. We don't actually test how much we should get disconnected,
as that's quite hard to do due to the reaper thread attempting to
connect as soon as it notices the connection going down.

As such, we simply print out how many disconnects we've encountered
and leave it up to human eyes (and brains) to discern if that's good
or bad.

Signed-off-by: Andreas Ericsson <ae@op5.se>