Avoid MSG_NOSIGNAL and ignore SIGPIPE globally instead
Solaris (among others) doesn't support MSG_NOSIGNAL, so we avoid it
and ignore SIGPIPE globally instead. If we don't, we might end up
dumping core when trying to read from disconnected nodes (which we
shouldn't do, but we might). Since the error path is quite simple
for such cases (disconnect the node and reconnect later) we can
simply avoid the entire problem by ignoring the signal if we get
one. This means that errno gets set to EPIPE in case a SIGPIPE was
supposed to be generated, and we already handle that case properly.

Signed-off-by: Andreas Ericsson <ae@op5.se>