import: Default to incremental mode when running imports
It doesn't make sense to do a non-incremental import on an existing
database without truncating it, so if we're using a database but
neither --truncate-db nor --incremental is set we default to using
incremental mode.

Short of using gdb and modifying in-core parameters during runtime,
this leaves no way whatsoever of running a full import on an already
populated database.

Since incremental imports should be handled fairly quickly, this
should completely hide the problem discussed in issue #3811, which
was caused by the logging tables being locked up far too long for
the merlin daemon to work properly for a long time.

Signed-off-by: Andreas Ericsson <ae@op5.se>