Makefile: Janitor patch, getting rid of superfluous variables
Since the project name has now been firmly etched in stone in the minds
of all interested users, we can get rid of the shortcut to quickly
change the name of it and save ourselves some variables.

As a nice side-effect, we now also build nebtest as a primary target.

Signed-off-by: Andreas Ericsson <ae@op5.se>