module: Fix peer_id assignments
Previously we sorted the list with earliest started first, but then
we compared in reverse, causing us to jump out of the loop at the
first node started before us instead of at the last node started
after us.

This patch makes the not-so-obvious correction by altering the order
of the arguments passed to timeval_comp() when we're looping to find
our peer id.

This patch swaps the peer-id assignment order though, so that the
earliest started server gets the lowest peer-id. Given that it's
not uncommon to frequently restart nodes that have problems or
are being upgraded or otherwise fiddled with, that's probably a
Good Thing(tm) anyway.

As a reminder to self, we also add a brief comment about adding a
binary search to the list of peers, and why we don't consider it
worth it for the smallish merlin networks we're expecting to see.

Signed-off-by: Andreas Ericsson <ae@op5.se>