Prevent crash in mon node status
After a monitor crash, mon node status would crash as well with a
DivisionByZero exception when trying to calculate the percentage of
hosts/services a node takes care of. This also applies to the
hypothetical situation where an installation actually has 0 hosts and/or
services.

Workaround by defining 0 out of 0 as 0%. This fixes bug #3953.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>