Plug a filedescriptor leak in config sha1 calculation
We used to leak the filedescriptors of the hashed files when getting
config hashes. For most installations, this isn't much of a problem,
but if someone has a vast amount of config files and a low value of
RLIMIT_NOFILE, this could quickly become a real issue. It's still an
issue with fewer config files, but the time it takes for it to become
a problem is a lot longer.

Signed-off-by: Andreas Ericsson <ae@op5.se>