module: Be more careful when cancelling and joining reaper thread
Since Nagios can call the cleanup() function, which de-initializes
all modules, without having called the modules initialization functions,
we must take care not to release resources we haven't allocated.

This patch takes care of that, assuming we never get a thread with id 0,
in which case we'll leak 256KiB of memory once and once only. It's still
better than crashing, and since the error-path where this happens pretty
much always ends with Nagios terminating, it's not as if we're missing
out all that much.

Signed-off-by: Andreas Ericsson <ae@op5.se>