A good deal of hackery trying to work around a numeric typing problem in the libdbi interface.
Discovered that fetching the value of "SELECT COUNT(*)" doesn't work
in libdbi unless you know the EXACT type of integer which the
underlying driver is using (correct sizeof and signedness). See:

http://www.mail-archive.com/libdbi-users@lists.sourceforge.net/msg00125.html

the new hackery works on MySQL but not on Sqlite, apparently because
the dbi/sqlite3 driver always returns a 0 value for this particular
case, even if we call every single one of the get-int variants.

Signed-off-by: Andreas Ericsson <ae@op5.se>