New UTF-8 approach
The old version tried to be Correct. The problem with being Correct is
that it breaks when users are doing incorrect things, such as storing
ISO-8859-1 data in their config. Instead, use ISO-8859-1 everywhere as a
sort of "random bytestring we know nothing about" type, and let ninja
deal with showing it correctly.

This works with the PHP importer. The Python importer is currently a bit
busted.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>