module: Don't let the reaper thread attempt to connect
If the reaper thread is allowed to connect to the daemon there's a
chance we run into a race-condition in the ipc connection api. This
could happen if the ipc connection has been lost, Nagios spawns an
event and the event handler thread fires off the connection attempt
at almost the same time as the reaper thread does it. The daemon
will then disconnect the one that happened to connect first and
accept the second connection instead.

That in itself isn't much of a problem, but if we're supposed to
send paths (ie, force a re-import of objects.cache and status.log)
and the reaper thread wins the race to connect so that the reaper
thread is the one that gets disconnected, the connect handler will
stash the paths event in the binlog for the ipc.

Since the binlog api is only threadsafe on a per-binlog basis,
this could lead to memory corruption and double-free()'s that some
customers have experienced on an infrequent basis.

It could potentially also happen if the two threads simultaneously
start to empty the binlog and race each other to the free() call.

Hopefully this patch will resolve the problem, even though it means
we'll now potentially delay a bit between each reconnection attempt.

The longest time we'll delay is MERLIN_PULSE_INTERVAL seconds, which
is a hardcoded value set to 10 seconds. In practice, it's unlikely
that the full pulse interval will occur between them, since each
new event to be sent from module to daemon will also trigger a new
connection attempt.

Signed-off-by: Andreas Ericsson <ae@op5.se>